refactor (#15) : Fix project structure
