--- src/main/java/io/toyproject/bookjeokclone/api/BookApi.java
@@ -1,21 +1,0 @@
-package io.toyproject.bookjeokclone.api;
-
-import io.toyproject.bookjeokclone.model.dto.BookRequest;
-import io.toyproject.bookjeokclone.model.entity.Book;
-import io.toyproject.bookjeokclone.service.BookService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/book")
-public class BookApi {
-
-    private final BookService bookService;
-
-    @PostMapping
-    public Book create(@RequestBody BookRequest request) {
-        return bookService.create(request);
-    }
-
-}
