+++ src/main/java/io/toyproject/bookjeokclone/board/model/dto/BoardRequest.java
@@ -1,0 +1,19 @@
+package io.toyproject.bookjeokclone.board.model.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class BoardRequest {
+
+
+    private String title;
+
+    private String writer;
+
+    private String content;
+}
