+++ src/main/java/io/toyproject/bookjeokclone/book/model/dto/AladinApiRequest.java
@@ -1,0 +1,7 @@
+package io.toyproject.bookjeokclone.book.model.dto;
+
+
+public class AladinApiRequest {
+    private String ttbKey;
+    private String query;
+}
