+++ src/main/java/io/toyproject/bookjeokclone/book/model/dto/AladinApiResponse.java
@@ -1,0 +1,17 @@
+package io.toyproject.bookjeokclone.book.model.dto;
+
+public class AladinApiResponse {
+    private int version;
+    private String tile;
+    private String link;
+    private String pubDate;
+    private int totalResults;
+    private int startIndex;
+    private int itemsPerPage;
+    private String query;
+    private int searchCategoryId;
+    private String searchCategoryName;
+
+    private AladinItem item;
+
+}
