+++ src/main/java/io/toyproject/bookjeokclone/book/model/dto/AladinItem.java
@@ -1,0 +1,15 @@
+package io.toyproject.bookjeokclone.book.model.dto;
+
+import java.util.Date;
+
+public class AladinItem {
+    private String title;
+    private String link;
+    private String author;
+    private String description;
+    private String isbn;
+    private String cover;
+    private String publisher;
+
+
+}
