+++ src/main/java/io/toyproject/bookjeokclone/book/model/dto/BookRequest.java
@@ -1,0 +1,22 @@
+package io.toyproject.bookjeokclone.book.model.dto;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class BookRequest {
+
+    private String title;
+
+    private String writer;
+
+    private String detail;
+
+    private String publisher;
+
+    private String isbn;
+
+    private Integer page;
+
+}
