+++ src/main/java/io/toyproject/bookjeokclone/book/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package io.toyproject.bookjeokclone.book.repository;
+
+import io.toyproject.bookjeokclone.book.model.entity.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
