+++ src/main/java/io/toyproject/bookjeokclone/book/service/AladinApiService.java
@@ -1,0 +1,8 @@
+package io.toyproject.bookjeokclone.book.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AladinApiService {
+
+}
