+++ src/main/java/io/toyproject/bookjeokclone/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package io.toyproject.bookjeokclone.member.repository;
+
+import io.toyproject.bookjeokclone.member.model.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
