--- src/main/java/io/toyproject/bookjeokclone/model/dto/AladinApiRequest.java
@@ -1,7 +1,0 @@
-package io.toyproject.bookjeokclone.model.dto;
-
-
-public class AladinApiRequest {
-    private String ttbKey;
-    private String query;
-}
