--- src/main/java/io/toyproject/bookjeokclone/model/dto/MemberRequest.java
@@ -1,14 +1,0 @@
-package io.toyproject.bookjeokclone.model.dto;
-
-import lombok.Data;
-
-@Data
-public class MemberRequest {
-
-    private String username;
-
-    private String email;
-
-    private String password;
-
-}
