--- src/main/java/io/toyproject/bookjeokclone/model/entity/Member.java
@@ -1,73 +1,0 @@
-package io.toyproject.bookjeokclone.model.entity;
-
-import io.toyproject.bookjeokclone.model.dto.MemberRequest;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "member")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@EntityListeners({AuditingEntityListener.class})
-public class Member {
-
-    @GenericGenerator(
-            name = "memberSequenceGenerator",
-            strategy = "org.hibernate.id.enhanced.SequenceStyleGenerator",
-            parameters = {
-                    @org.hibernate.annotations.Parameter(name = "sequence_name", value = "SEQ_MEMBER"),
-                    @org.hibernate.annotations.Parameter(name = "initial_value", value = "1"),
-                    @org.hibernate.annotations.Parameter(name = "increment_size", value = "1")
-            }
-    )
-    @Id
-    @GeneratedValue(generator = "memberSequenceGenerator")
-    private Long id;
-
-    @Column(name = "username", nullable = false)
-    private String username;
-
-    @Column(name = "email", nullable = false)
-    private String email;
-
-    @Column(name = "password", nullable = false)
-    private String password;
-
-    @CreatedDate
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(name = "updated_at", nullable = true)
-    private LocalDateTime updatedAt;
-
-    public static Member create(MemberRequest request) {
-        return Member.builder()
-                .username(request.getUsername())
-                .email(request.getEmail())
-                .password(request.getPassword())
-                .build();
-    }
-
-    @Builder
-    private Member(String username, String email, String password) {
-        this.username = username;
-        this.email = email;
-        this.password = password;
-    }
-}
