--- src/main/java/io/toyproject/bookjeokclone/repository/BookRepository.java
@@ -1,7 +1,0 @@
-package io.toyproject.bookjeokclone.repository;
-
-import io.toyproject.bookjeokclone.model.entity.Book;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookRepository extends JpaRepository<Book, Long> {
-}
