--- src/main/java/io/toyproject/bookjeokclone/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package io.toyproject.bookjeokclone.repository;
-
-import io.toyproject.bookjeokclone.model.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
